refactor(deno): split the wrapper into one module per domain
src/deno.ts held the settings base, every command class and the task
object in one 773-line file. Guideline 8 asks for one domain per file,
and the file is about to grow by sixteen more commands, so it splits
first and gains them second.

Each class moves verbatim into the module for its domain — settings.ts
for the two bases, execution.ts for run and task, testing.ts for test
and coverage, quality.ts for check, fmt, lint and doc, dependencies.ts
for cache, install and publish — leaving deno.ts as the assembler that
imports them and exposes DenoTasks.

No argv, no validation and no signature changes travel with the move:
the existing suite passes as written, and the only edits to it are its
import path, which now points at the package entry point rather than at
one internal module, and one added reach assertion for DenoTasks.doc,
which was the single task the suite never drove to tool resolution.
